requirements.txt
# Python dependencies for AB1 Sequencing QC Pipeline
# Core dependencies
biopython>=1.83
pandas>=2.0
matplotlib>=3.0
numpy>=1.20.0
ipykernel>=6.0
ipython>=7.0
click>=8.0
pyyaml>=6.0

# Development dependencies (optional)
pytest>=7.0
pytest-cov>=4.0
black>=22.0
flake8>=5.0
mypy>=1.0

# Documentation dependencies (optional)
sphinx>=5.0
furo>=2024.1.29
sphinx-copybutton>=0.5.0
myst-parser>=2.0.0
sphinxcontrib-mermaid>=1.0.0